requirements.txt
# Core Discord Bot Dependencies
discord.py==2.3.2
python-dotenv==1.0.0

# AI Integration
groq==0.30.0
openai==1.57.0

# Google Search Integration  
google-api-python-client==2.108.0

# Web Requests (for Perplexity API)
aiohttp==3.9.1

# Crafting System (local module)
# dune_crafting.py is included in the project